refactor(pedagogy): extract prompt_pack tests to sibling file
`prompt_pack.rs` was 2072 lines, dominated by a ~1180-line
`#[cfg(test)] mod tests`. Move that module verbatim into a sibling
`prompt_pack/tests.rs` (Rust-2018 `foo.rs` + `foo/` form), leaving the
parent at 894 lines. As a child module, the tests still see every
private item of the parent via `use super::*` — zero visibility churn.

The test-support helper `reset_preview_warn_once_for_test` stays in the
parent (`#[cfg(test)] pub(super)`); descendant modules read ancestor
items regardless of visibility, so no signature change was needed.

Verbatim relocation: `cargo fmt` normalised the de-indented wrapping
whitespace while leaving the `r#"..."#` TOML fixtures intact. The 69
prompt_pack tests pass identically under their new
`prompt_pack::tests::*` paths — the existing tests are the regression
guard. No behaviour change.
